migration of few more files to ts
